Increase the limit of max open files to 64k
If current rlimit (OS soft limit) for max open file descriptors is less than 64k, bump
it to 64k (or to hard limit, if hard limit is smaller).
It should be enough, as normally the limit is 1k.

It is still possible to set an even higher number manually using
'ulimit' command before launching the node.

With this change, we can now increase the limit of max open files in
RocksDB configuration.