Remove db-cfg-max-open-files config option
The default RocksDB setting for max_open_files is -1 (unlimited) and it
is recommended to leave this setting for performance reasons - RocksDB can then
skip expensive checks during runtime. There are more hints to leave the
-1 setting in internal RocksDB code.

Commit 05477af7 increased process file descriptor limit to 64k so we should
never hit the OS limit of open files. We can also control the number of
files RocksDB creates by knobs such as target_file_size_base/multiplier.

In case the limit is ever hit and open() calls fail, RocksDB propagates
the error to the caller so it would be a matter of handling the error
correctly in the node.