Simplify RocksDB options and add a 128MB Block Cache
Using a 128MB Block Cache instead of the default 8MB results in huge
gains in terms of db.get tail latency - ~50% lower P95 and P99 latency.
Increase in RSS memory usage was not observed.
The 128MB value was taken from official RocksDB recommendations.

Below are some RocksDB metrics after 30+hr use with this change:
- block cache hit to miss ratio - 3.05
- index hit to miss ratio - 393.5
- data hit to miss ratio - 2.39
- memtable hit to miss ratio - 5.5

The Block Cache is shared among all column families. It is safe to use
it that way even with multiple threads. RocksDB C++ code uses a mutex
for the cache internally. Rust-rocksdb bindings do not do anything do
the Cache pointer except call create/destroy on it.

One caveat is that the Cache object must live at least as long as the DB
object, which is why it is allocated in main().