Use immutable OrdMap instead of BTreeMap for trees
Cloning BTreeMaps is very expensive and since it's done a lot in
get_entry() the allocation overhead is very significant. Using immutable
OrdMap reduces this overhead immensely with 2-4x less time spent on
allocations and 2-4x less memory being allocated during actors test.

Note that im::OrdMap uses Arc and now Arc::new() allocation is the
biggest hotspot. In the future, we need to refactor compute_new_root to
remove tree cloning so that im:: will not be necessary.