Implement get() from staging
Reimplement get() to work with the new version. On get() call, build the
staging area from all previous actions, then get the value.
Caveats:
- this can be slow if the list of actions is very large.
- has a side effect - actions list is cleared so we cannot backtrack
after get() is called, before the next commit().

More tests need to be added for get() as it is not used in the normal
flow.

Fix tests and bugs in remove, copy and checkout